Move consensus error code definition to result-code.js.
