+++ consensus/constants.js
@@ -22,69 +22,9 @@ const ValidatorOffenseTypes = {
   INVALID_PROPOSAL: 'INVALID_PROPOSAL',
 }
 
-const ConsensusErrorCode = {
-  PARSING_PROPOSAL_BLOCK_FAILURE: 90001,
-  NOT_EXTENDING_LNC_ERROR: 90002,
-  TEMP_DB_CREATION_FAILURE: 90003,
-  INVALID_CONSENSUS_TX: 90100,
-  PROPOSER_MISMATCH: 90103,
-  BLOCK_HASH_MISMATCH: 90104,
-  MISSING_PREV_BLOCK: 90200,
-  MISSING_DB_FOR_PREV_BLOCK: 90201,
-  ILL_FORMED_BLOCK: 90300,
-  INVALID_GENESIS_BLOCK: 90301,
-  INVALID_CHAIN: 90302,
-  INVALID_EPOCH: 90303,
-  INVALID_VALIDATORS_SIZE: 90304,
-  INVALID_PROPOSER: 90305,
-  EXECUTING_LAST_VOTES_FAILURE: 90306,
-  INVALID_LAST_VOTES_STAKES: 90307,
-  MISSING_PROPOSAL_IN_LAST_VOTES: 90308,
-  OFFENSES_EVIDENCE_MISMATCH: 90309,
-  INVALID_OFFENSE_TYPE: 90310,
-  INVALID_EVIDENCE_VOTES_STAKES: 90311,
-  EXECUTING_EVIDENCE_VOTES_FAILURE: 90312,
-  INVALID_OFFENSE_COUNTS: 90313,
-  EXECUTING_TX_FAILURE: 90314,
-  INVALID_RECEIPTS: 90315,
-  INVALID_GAS_AMOUNT_TOTAL: 90316,
-  INVALID_GAS_COST_TOTAL: 90317,
-  INVALID_STATE_PROOF_HASH: 90318,
-  ILL_FORMED_PROPOSAL_TX: 90319,
-  EXECUTING_PROPOSAL_FAILURE: 90320,
-  ADDING_TO_BLOCK_POOL_FAILURE: 90400,
-}
-
-const ConsensusErrorCodesToVoteAgainst = new Set([
-  ConsensusErrorCode.ADDING_TO_BLOCK_POOL_FAILURE,
-  ConsensusErrorCode.INVALID_EPOCH,
-  ConsensusErrorCode.MISSING_PROPOSAL_IN_LAST_VOTES,
-  ConsensusErrorCode.EXECUTING_LAST_VOTES_FAILURE,
-  ConsensusErrorCode.INVALID_LAST_VOTES_STAKES,
-  ConsensusErrorCode.INVALID_PROPOSER,
-  ConsensusErrorCode.BLOCK_HASH_MISMATCH,
-  ConsensusErrorCode.INVALID_VALIDATORS_SIZE,
-  ConsensusErrorCode.ILL_FORMED_BLOCK,
-  ConsensusErrorCode.OFFENSES_EVIDENCE_MISMATCH,
-  ConsensusErrorCode.INVALID_OFFENSE_TYPE,
-  ConsensusErrorCode.INVALID_EVIDENCE_VOTES_STAKES,
-  ConsensusErrorCode.EXECUTING_EVIDENCE_VOTES_FAILURE,
-  ConsensusErrorCode.INVALID_OFFENSE_COUNTS,
-  ConsensusErrorCode.EXECUTING_TX_FAILURE,
-  ConsensusErrorCode.INVALID_RECEIPTS,
-  ConsensusErrorCode.INVALID_GAS_AMOUNT_TOTAL,
-  ConsensusErrorCode.INVALID_GAS_COST_TOTAL,
-  ConsensusErrorCode.INVALID_CONSENSUS_TX,
-  ConsensusErrorCode.ILL_FORMED_PROPOSAL_TX,
-  ConsensusErrorCode.EXECUTING_PROPOSAL_FAILURE,
-  ConsensusErrorCode.INVALID_STATE_PROOF_HASH,
-]);
-
 module.exports = {
   ConsensusMessageTypes,
   ConsensusConsts,
   ConsensusStates,
   ValidatorOffenseTypes,
-  ConsensusErrorCode,
-  ConsensusErrorCodesToVoteAgainst,
 }
